+++ packages/core/rest/con.auth.routes.js
@@ -4,7 +4,7 @@
  * @import { ApiQuery } from '../api/types.api.query.js' 
  */
 import { App } from '../index.js';
-import { Polka } from '../polka/index.js'
+import { Polka } from './polka/index.js'
 import { authorize_admin } from './con.auth.middle.js';
 import { parse_query } from '../api/utils.query.js';
 import { create_routes as idp_routes } from './con.auth.identity-providers.routes.js';
