+++ packages/core/rest/con.customers.routes.js
@@ -3,7 +3,7 @@
  * @import { CustomerType, OrderData } from "../api/types.api.js"; 
  * @import { ApiQuery } from "../api/types.api.query.js"; 
  */
-import { Polka } from '../polka/index.js'
+import { Polka } from './polka/index.js'
 import { assert } from '../api/utils.func.js'
 import { parse_auth_user, roles_guard } from './con.auth.middle.js'
 import { parse_query } from '../api/utils.query.js'
