+++ packages/core/rest/con.posts.routes.js
@@ -4,7 +4,7 @@
  * @import { PostType } from '../api/types.api.js' 
  */
 import { App } from '../index.js';
-import { Polka } from '../polka/index.js'
+import { Polka } from './polka/index.js'
 import { assert } from '../api/utils.func.js'
 import { authorize_by_roles } from './con.auth.middle.js'
 import { parse_query } from '../api/utils.query.js'
