+++ packages/core/rest/con.products.routes.js
@@ -5,7 +5,7 @@
  * @import { RegularGetOptions } from '../database/types.public.js' 
  */
 import { App } from '../index.js';
-import { Polka } from '../polka/index.js'
+import { Polka } from './polka/index.js'
 import { assert } from '../api/utils.func.js'
 import { authorize_by_roles } from './con.auth.middle.js'
 import { parse_expand, parse_query } from '../api/utils.query.js'
