+++ packages/core/rest/con.search.routes.js
@@ -2,7 +2,7 @@
  * @import { ApiPolka } from './types.public.js' 
  */
 import { App } from '../index.js';
-import { Polka } from '../polka/index.js'
+import { Polka } from './polka/index.js'
 import { parse_query } from '../api/utils.query.js'
 import { is_admin, parse_auth_user } from './con.auth.middle.js';
