+++ packages/core/rest/con.similarity-search.routes.js
@@ -6,7 +6,7 @@
  */
 
 import { App } from '../index.js';
-import { Polka } from '../polka/index.js'
+import { Polka } from './polka/index.js'
 import { parse_list_from_string, parse_number_from_string } from '../api/utils.query.js'
 import { is_admin, parse_auth_user } from './con.auth.middle.js';
