+++ packages/core/rest/index.js
@@ -1,10 +1,11 @@
 /** 
  * @import { ApiPolka, ApiRequest, ApiResponse } from './types.public.js' 
- * @import { CORSOptions } from '../polka/cors.js' 
+ * @import { CORSOptions } from './polka/cors.js' 
  */
 import { App } from '../index.js';
-import { json } from "../polka/body-parse.js";
-import { Polka } from "../polka/index.js";
+import { json } from "./polka/body-parse.js";
+import { cors } from "./polka/cors.js";
+import { Polka } from "./polka/index.js";
 import { create_routes as create_auth_route } from "./con.auth.routes.js";
 import { create_routes as create_tags_route } from "./con.tags.routes.js";
 import { create_routes as create_templates_route } from "./con.templates.routes.js";
@@ -30,7 +31,6 @@ import { create_routes as create_search_route } from "./con.search.routes.js";
 import { create_routes as create_ai_route } from "./con.ai.routes.js";
 import { create_routes as create_similarity_search_route } from "./con.similarity-search.routes.js";
 import { create_routes as create_emails_route } from "./con.emails.routes.js";
-import { cors } from "../polka/cors.js";
 
 
 /**
