+++ packages/core/rest/polka/body-parse.js
@@ -1,13 +1,11 @@
+/**
+ * @import { VPolkaRequest, VPolkaResponse, Middleware, PolkaOptions, IPolka } from './public.js'
+ */
 const CONTENT_TYPE = 'content-type';
 const CONTENT_TYPE_JSON = 'application/json';
 const CONTENT_TYPE_TEXT = 'text/plain';
 const CONTENT_TYPE_OCTET_STREAM = 'application/octet-stream';
 const CONTENT_TYPE_FORM_URL_ENCODED = 'application/x-www-form-urlencoded';
-
-/**
- * @typedef {import("./public.d.ts").VPolkaRequest} VPolkaRequest
- * @typedef {import("./public.d.ts").VPolkaResponse} VPolkaResponse
- */
 
 /**
  * 
@@ -19,6 +17,7 @@ const is_type = (req, match_type) => {
 }
 
 /**
+ * @description Parse JSON body
  * @param {boolean} [keep_raw=true]
  */
 export const json = (keep_raw=true) => {
@@ -79,9 +78,8 @@ export const raw_buffer = () => {
 }
 
 /**
- * 
+ * @description Parse form data into FormData or Object
  * @param {boolean} parse_to_object instead of FormData, parse into regular Object dictionary
- * @returns 
  */
 export const urlencoded = (parse_to_object=true) => {
   /**
