+++ packages/core/rest/polka/index.js
@@ -1,13 +1,9 @@
-import { StorecraftError } from '../api/utils.func.js';
+/**
+ * @import { VPolkaRequest, VPolkaResponse, Middleware, PolkaOptions, IPolka } from './public.js'
+ */
+import { StorecraftError } from '../../api/utils.func.js';
 import { STATUS_CODES } from './codes.js';
 import { Trouter } from './trouter/index.js';
-
-/**
- * @typedef {import('./public.d.ts').VPolkaRequest} VPolkaRequest 
- * @typedef {import('./public.d.ts').VPolkaResponse} VPolkaResponse 
- * @typedef {import('./public.d.ts').IPolka} IPolka3
- * 
- */
 
 const extract_code = e => {
   const parsed = parseInt(e?.cause)
@@ -50,30 +46,23 @@ export const onError = async (error, req, res) => {
 /**
  * @template {VPolkaRequest} Req
  * @template {VPolkaResponse} Res
- * @param {import('./public.d.ts').Middleware<Req, Res>} fn 
- * @returns {import('./public.d.ts').Middleware<Req, Res>}
+ * @param {Middleware<Req, Res>} fn 
+ * @returns {Middleware<Req, Res>}
  */
 const mount = fn => fn instanceof Polka ? fn.attach : fn;
 
 /**
- * @template {VPolkaRequest} Req
- * @template {VPolkaResponse} Res
- * @typedef {import('./public.d.ts').IPolka<Req, Res>} IPolka
- * 
- */
-
-/**
  * 
  * @template {VPolkaRequest} Req
  * @template {VPolkaResponse} Res
- * @extends {Trouter<import('./public.d.ts').Middleware<Req, Res>>}
+ * @extends {Trouter<Middleware<Req, Res>>}
  * @implements {IPolka<Req, Res>}
  */
 export class Polka extends Trouter {
 
   /**
    * 
-   * @param {import('./public.d.ts').PolkaOptions<Req, Res>} opts 
+   * @param {PolkaOptions<Req, Res>} opts 
    */
   constructor(opts = {}) {
     super();
@@ -87,9 +76,9 @@ export class Polka extends Trouter {
   }
 
   /**
-   * @typedef {(RegExp | string | IPolka<Req, Res> | import('./public.d.ts').Middleware<Req, Res>)} Every
+   * @typedef {(RegExp | string | IPolka<Req, Res> | Middleware<Req, Res>)} Every
    * @param {Every} base 
-   * @param  {...(Polka | import('./public.d.ts').Middleware<Req, Res>)} fns 
+   * @param  {...(Polka | Middleware<Req, Res>)} fns 
    * @returns 
    */
   use(base, ...fns) {
